Refactor journal feature with a custom widget
This commit refactors the journal feature by replacing the old implementation with a new custom widget.

The new implementation is more modular and reusable. It consists of two new widgets:
- `CustomJournalWidget`: a stateful widget that manages the state of the journal. It can be used to create, edit, and view journal entries.
- `JournalListView`: a widget that displays the list of journal entries.

The new implementation uses `flutter_quill` to provide a rich text editor. The journal entries are saved to a local SQLite database.

The old journal files have been deleted.
