+++ pubspec.yaml
@@ -63,6 +63,8 @@ dependencies:
   url_launcher: ^6.3.1
   device_info_plus: ^11.5.0
   flutter_quill: ^11.4.2
+  flutter_keyboard_visibility: ^6.0.0
+  flutter_quill_localizations: ^11.4.2
 
 dev_dependencies:
   flutter_test:
